회원별 결제 및 예약 내역 조회 API 구현

+++ src/main/java/com/daview/service/ReservationServiceImpl.java
@@ -22,7 +22,7 @@ public class ReservationServiceImpl implements ReservationService {
 	}
 
 	@Override
-	public List<ReservationDTO> selectReservationById(int memberId) {
+	public List<ReservationDTO> selectReservationById(Long memberId) {
 		return reservationMapper.selectReservationById(memberId);
 	}
